Updated header navigation bar. Updating Home view.
