+++ VolunteerOrganizer/src/main/resources/templates/fragments.html
@@ -12,7 +12,7 @@
         <div>
             <a th:href="@{/home}">Home</a>
             <a th:href="@{/create}">Create Volunteer Opportunity</a>
-            <a th:href="@{/search}">Search Opportunity</a>
+            <a th:href="@{/logout}">Logout</a></li>
         </div>
     <h3 style="color:brown" th:text ="${title}"></h3><br>
     </div>
